Gadisa.rate crud fix (#143)
* fix(gadisa.rate-crud-fix): fix BlogApp2.sln file

* fix(gadisa.rate-crud-fix): fix deletecommandhandler  error

* fix(gadisa.rate-crud-fix): rename Rate folder to Ratetest in test project
