+++ Backend/starter_project_dotnetforce7/BlogApp2/BlogApp.Application.UnitTest/Ratetest/Command/DeleteRateCommandHandlerTest.cs
@@ -1,7 +1,7 @@
 ﻿using AutoMapper;
 using BlogApp.Application.Contracts.Persistence;
-using BlogApp.Application.Features.Rates.CQRS.Commands;
 using BlogApp.Application.Features.Rates.CQRS.Handlers;
+using BlogApp.Application.Features.Rates.CQRS.Commands;
 using BlogApp.Application.Features.Rates.DTOs;
 using BlogApp.Application.Profiles;
 using BlogApp.Application.Responses;
@@ -11,21 +11,21 @@ using Moq;
 using Shouldly;
 using System;
 using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace BlogApp.Application.UnitTest.Rates.Command
+namespace BlogApp.Application.UnitTest.Ratetest.Command
 {
-    public class CreateRateCommandHandlerTest
+    public class DeleteRateCommandHandlerTest
     {
 
         private readonly IMapper _mapper;
         private readonly Mock<IUnitOfWork> _mockRepo;
+        private int _id { get; set; }
+        private readonly DeleteRateCommandHandler _handler;
         private readonly CreateRateDto _rateDto;
-        private readonly CreateRateCommandHandler _handler;
-        public CreateRateCommandHandlerTest()
+        public DeleteRateCommandHandlerTest()
         {
             _mockRepo = MockUnitOfWork.GetUnitOfWork();
             var mapperConfig = new MapperConfiguration(c =>
@@ -33,47 +33,46 @@ namespace BlogApp.Application.UnitTest.Rates.Command
                 c.AddProfile<MappingProfile>();
             });
             _mapper = mapperConfig.CreateMapper();
-
             _rateDto = new CreateRateDto
             {
-                RateNo=5,
-                RaterId=3,
+                RateNo = 5,
+                RaterId = 3,
                 BlogId = 4
             };
+            _id = 1;
 
-            _handler = new CreateRateCommandHandler(_mockRepo.Object, _mapper);
+            _handler = new DeleteRateCommandHandler(_mockRepo.Object, _mapper);
 
         }
 
 
         [Fact]
-        public async Task CreateRate()
+        public async Task DeleteRate()
         {
-            var result = await _handler.Handle(new CreateRateCommand() { RateDto = _rateDto }, CancellationToken.None);
-            result.ShouldBeOfType<Result<int>>();
+            /* var create_result = await _handler.Handle(new CreateRateCommand(){ RateDto = _rateDto  }, CancellationToken.None);*/
+
+            var result = await _handler.Handle(new DeleteRateCommand() { Id = _id }, CancellationToken.None);
+            result.ShouldBeOfType<Result<Unit>>();
             result.Success.ShouldBeTrue();
 
             var rates = await _mockRepo.Object.RateRepository.GetAll();
-            rates.Count.ShouldBe(3);
-
+            rates.Count().ShouldBe(1);
         }
 
         [Fact]
-        public async Task InvalidRate_Added()
+        public async Task Delete_Rate_Doesnt_Exist()
         {
 
-            _rateDto.RateNo = -1;
-            var result = await _handler.Handle(new CreateRateCommand() { RateDto = _rateDto }, CancellationToken.None);
-            result.ShouldBeOfType<Result<int>>();
-            result.Success.ShouldBeFalse();
-            result.Errors.ShouldNotBeEmpty();
+            _id  = 0;
+            var result = await _handler.Handle(new DeleteRateCommand() { Id = _id }, CancellationToken.None);
+            result.ShouldBe(null);
+        
             var rates = await _mockRepo.Object.RateRepository.GetAll();
             rates.Count.ShouldBe(2);
 
         }
     }
 }
-
