+++ Backend/starter_project_dotnetforce7/BlogApp2/BlogApp.Application.UnitTest/Ratetest/Query/GetRateDetailQueryHandlerTest.cs
@@ -15,7 +15,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace BlogApp.Application.UnitTest.Rates.Query
+namespace BlogApp.Application.UnitTest.Ratetest.Query
 {
     public class GetRateDetailQueryHandlerTest
     {
